backend/requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
requests
azure-storage-blob
pyodbc